Extract WebHelpMenu.java from TripleAMenuBar
